feat(time-picker): add new version (#341)
* feat(time-picker): add new version

* fix(input): fix input autocomplete props
